wip refactor testing
