+++ StandardLibrary/Sources/LibC.hylo
@@ -1,0 +1,9 @@
+/// Allocates `size` bytes of uninitialized storage.
+@c_ffi("c_malloc_indirect")
+public fun malloc(size: Int) -> Pointer<Never> // TODO use MemoryAddress
+
+/// Deallocates the block of memory pointed at by `pointer`.
+/// 
+/// - Requires: there is a live heap allocation starting at `address`.
+@c_ffi("c_free_indirect")
+public fun unsafe_free(pointer: Pointer<Never>) // TODO use MemoryAddress
