+++ Tests/CompilerTests/positive/c-ffi-calls.package/package.json
@@ -1,0 +1,1 @@
+{ "options": ["exit-status:6"] }
